+++ app/src/main/java/com/afrofx/code/anjesgf/models/RegistoVendaModel.java
@@ -5,60 +5,60 @@ package com.afrofx.code.anjesgf.models;
  */
 
 public class RegistoVendaModel {
-    private int id_user, id_registo_operacao, id_cliente, id_registo_venda, venda_estado;
+    private long id_user, id_registo_operacao, id_cliente, id_registo_venda, venda_estado;
     private String venda_data_expiracao;
     private double venda_desconto;
 
-    public RegistoVendaModel(int id_user, int id_registo_operacao, int venda_estado) {
+    public RegistoVendaModel(long id_user, long id_registo_operacao, long venda_estado) {
         this.id_user = id_user;
         this.id_registo_operacao = id_registo_operacao;
         this.venda_estado = venda_estado;
     }
 
-    public RegistoVendaModel(int id_user, int id_registo_operacao, int venda_estado, int id_cliente) {
+    public RegistoVendaModel(long id_user, long id_registo_operacao, long venda_estado, long id_cliente) {
         this.id_user = id_user;
         this.id_registo_operacao = id_registo_operacao;
         this.venda_estado = venda_estado;
         this.id_cliente = id_cliente;
     }
 
-    public int getId_user() {
+    public long getId_user() {
         return id_user;
     }
 
-    public void setId_user(int id_user) {
+    public void setId_user(long id_user) {
         this.id_user = id_user;
     }
 
-    public int getId_registo_operacao() {
+    public long getId_registo_operacao() {
         return id_registo_operacao;
     }
 
-    public void setId_registo_operacao(int id_registo_operacao) {
+    public void setId_registo_operacao(long id_registo_operacao) {
         this.id_registo_operacao = id_registo_operacao;
     }
 
-    public int getId_cliente() {
+    public long getId_cliente() {
         return id_cliente;
     }
 
-    public void setId_cliente(int id_cliente) {
+    public void setId_cliente(long id_cliente) {
         this.id_cliente = id_cliente;
     }
 
-    public int getId_registo_venda() {
+    public long getId_registo_venda() {
         return id_registo_venda;
     }
 
-    public void setId_registo_venda(int id_registo_venda) {
+    public void setId_registo_venda(long id_registo_venda) {
         this.id_registo_venda = id_registo_venda;
     }
 
-    public int getVenda_estado() {
+    public long getVenda_estado() {
         return venda_estado;
     }
 
-    public void setVenda_estado(int venda_estado) {
+    public void setVenda_estado(long venda_estado) {
         this.venda_estado = venda_estado;
     }
